pyproject.toml
[project]
name = "library"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "patool>=4.0.1",
    "reformer-pytorch>=1.4.4",
    "scikit-learn>=1.7.2",
    "sktime>=0.39.0",
    "tqdm>=4.67.1",
]
